+++ include/qpdf/Pl_Buffer.hh
@@ -32,7 +32,7 @@ class Pl_Buffer: public Pipeline
     QPDF_DLL
     virtual ~Pl_Buffer();
     QPDF_DLL
-    virtual void write(unsigned char*, int);
+    virtual void write(unsigned char*, size_t);
     QPDF_DLL
     virtual void finish();
